change places for config to be in separate file due to conflicts with tests in costumercontrollertest. Should workn now.
